editoast: reorganize rolling stock modules
