+++ editoast/src/views/rolling_stock/form.rs
